improve getMessageHTML signature, improve telegram message crop
Previously, the telegram message was stripped of its HTML,
which was unintended. After that change, it would be
stripped of HTML formatting only if the characters
limit is exceeded. Otherwise, we need to do sophisticated
HTML parsing to find the tag, which would be left without the
closing part when trimming the message.