upload file to telegram from FS (instead of streaming it)
Potential resolution for #78 as the problem might be caused
by slow upload to Telegram, which causes the connection to the
server we're reading the file from to drop. In the new setup,
the application downloads the file first. Then uploads it to Telegram
(and calculates the track duration) at the speed
which is comfortable for Telegram, without making the uploading
server wait.