fix large file upload fallback
Fallback is supposed to kick in the situation where the uploaded file
is too big for Telegram and send a text message instead of audio
in that case. It stopped working likely after library update and
resulting error message change: this commit makes sure that
the error is parsed correctly in the new conditions.